feat: Complete ParticleManager.js splitting (Task 5)
- Refactored ParticleManager.js to lightweight main controller (2,728→1,035 words)
- Delegated rendering operations to ParticleRenderer
- Delegated lifecycle management to ParticleLifecycleManager
- Removed individual drawing methods (delegated to renderer)
- Removed physics/appearance update methods (delegated to lifecycle)
- Maintained all public APIs through delegation
- All components now under 2,500-word MCP compatibility limit

Task 5 Complete: ParticleManager.js split into 3 components
✅ ParticleRenderer.js (1,152 words) - rendering & drawing
✅ ParticleLifecycleManager.js (2,051 words) - lifecycle & memory
✅ ParticleManager.js (1,035 words) - main controller
